qa/docker/requirements.txt
langchain==0.0.209
openai==0.27.1
pinecone_client==2.2.1
psycopg2==2.9.5
psycopg2_binary==2.9.5
python-dotenv==1.0.0
retrying==1.3.4
sendgrid==6.10.0
spacy==3.5.0
torch==2.0.0
tqdm==4.64.1
transformers==4.27.4
